Feedback for 1/06 review:
Hi Mike, here are some changes (I'm not quite sure I've done the correct formatting)
for some verbiage, error types, and a few examples.

I also have a few specific notes that aren't quite captured here:
- There are additional API calls for zone import/exports that are not present in these docs
I've noted the necessary changes in a note on `dns_with_rax_roles-2.wadl` on `L751`
- I'm not sure if my changing of the `3.12. Faults` section requires any other changes.
I know there are quite a few places that reference similar information.

A quick clarification as well. The V2 DNS API has almost nothing in common with the v1 API,
apart from a few similar terms. The main differences to keep in mind are:
- There is no XML anywhere in the v2 API, I deleted a few XML examples, but there should be
none because the API does not support XML.
- The v1 API had a distinct idea of "subdomains", these being resources you could query in the API
that were logically attached to a parent. In the v2 API subzones exist, but are implicit. There is no
explicit calling out of the fact that a zone is a subzone of another. I changed a few references
in these docs, but there may be more.

I hope this helps, and thank you so much for writing these docs!!!